requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
openai
python-dotenv
faiss-cpu
sentence-transformers
transformers
torch
easyocr
PyMuPDF
tqdm
numpy